backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
mysqlclient==2.2.0
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0
django-filter==23.3
gunicorn==21.2.0
whitenoise==6.6.0